+++ wpull/decompression.py
@@ -2,10 +2,40 @@
 '''Streaming decompressors.'''
 import zlib
 
-import tornado.util
+
+class SimpleGzipDecompressor(object):
+    """Streaming gzip decompressor.
+
+    The interface is like that of `zlib.decompressobj` (without some of the
+    optional arguments, but it understands gzip headers and checksums.
+    """
+    # This class taken from tornado.util.GzipDecompressor
+    # Copyright Facebook. License Apache License Version 2.0.
+    def __init__(self):
+        # Magic parameter makes zlib module understand gzip header
+        # http://stackoverflow.com/questions/1838699/how-can-i-decompress-a-gzip-stream-with-zlib
+        # This works on cpython and pypy, but not jython.
+        self.decompressobj = zlib.decompressobj(16 + zlib.MAX_WBITS)
+
+    def decompress(self, value):
+        """Decompress a chunk, returning newly-available data.
+
+        Some data may be buffered for later processing; `flush` must
+        be called when there is no more input data to ensure that
+        all data was processed.
+        """
+        return self.decompressobj.decompress(value)
+
+    def flush(self):
+        """Return any remaining buffered data not yet returned by decompress.
+
+        Also checks for errors such as truncated input.
+        No other methods may be called on this object after `flush`.
+        """
+        return self.decompressobj.flush()
 
 
-class GzipDecompressor(tornado.util.GzipDecompressor):
+class GzipDecompressor(SimpleGzipDecompressor):
     '''gzip decompressor with gzip header detection.
 
     This class checks if the stream starts with the 2 byte gzip magic number.
@@ -23,9 +53,11 @@ class GzipDecompressor(tornado.util.GzipDecompressor):
             else:
                 return value
         else:
-            # FIXME: don't assume that we receive 2 bytes or more on!
+            # XXX: gzip magic value is \x1f\x8b but data may come in as
+            # a single byte. The likelyhood of plaintext starting with \x1f is
+            # very low, right?
             self.checked = True
-            if value[:2] == b'\x1f\x8b':
+            if value[:1] == b'\x1f':
                 self.is_ok = True
                 return super().decompress(value)
             else:
@@ -39,7 +71,7 @@ class GzipDecompressor(tornado.util.GzipDecompressor):
             return b''
 
 
-class DeflateDecompressor(tornado.util.GzipDecompressor):
+class DeflateDecompressor(SimpleGzipDecompressor):
     '''zlib decompressor with raw deflate detection.
 
     This class doesn't do any special. It only tries regular zlib and then
@@ -76,7 +108,7 @@ def gzip_uncompress(data, truncated=False):
     Raises:
         zlib.error
     '''
-    decompressor = tornado.util.GzipDecompressor()
+    decompressor = SimpleGzipDecompressor()
     inflated_data = decompressor.decompress(data)
 
     if not truncated: